Replace all unsigned long wit size_t

+++ block-sha1/sha1.h
@@ -7,13 +7,13 @@
  */
 
 typedef struct {
-	unsigned long long size;
+	size_t long size;
 	unsigned int H[5];
 	unsigned int W[16];
 } blk_SHA_CTX;
 
 void blk_SHA1_Init(blk_SHA_CTX *ctx);
-void blk_SHA1_Update(blk_SHA_CTX *ctx, const void *dataIn, unsigned long len);
+void blk_SHA1_Update(blk_SHA_CTX *ctx, const void *dataIn, size_t len);
 void blk_SHA1_Final(unsigned char hashout[20], blk_SHA_CTX *ctx);
 
 #define platform_SHA_CTX	blk_SHA_CTX
